+++ PosOnLine/OOB/Documento/Agregar/Factura/Ficha.cs
@@ -121,7 +121,7 @@ namespace PosOnLine.OOB.Documento.Agregar.Factura
         public List<FichaMedida> Medidas { get; set; }
         //
         public decimal PorctBonoPorPagoDivisa { get; set; }
-        public int CantDivisaAplicaBonoPorPagoDivisa { get; set; }
+        public decimal CantDivisaAplicaBonoPorPagoDivisa { get; set; }
         public decimal MontoBonoPorPagoDivisa { get; set; }
         public decimal MontoBonoEnDivisaPorPagoDivisa { get; set; }
         public decimal MontoPorVueltoEnEfectivo { get; set; }
@@ -243,7 +243,7 @@ namespace PosOnLine.OOB.Documento.Agregar.Factura
             Medidas = new List<FichaMedida>();
             //
             PorctBonoPorPagoDivisa = 0m;
-            CantDivisaAplicaBonoPorPagoDivisa = 0;
+            CantDivisaAplicaBonoPorPagoDivisa = 0m;
             MontoBonoPorPagoDivisa = 0m;
             MontoBonoEnDivisaPorPagoDivisa = 0m;
             MontoPorVueltoEnEfectivo = 0m;
