+++ PosOnLine/Src/Pago/Procesar/dataRecolectar.cs
@@ -12,7 +12,7 @@ namespace PosOnLine.Src.Pago.Procesar
     {
 
         public decimal PorctBonoPorPagoDivisa { get; set; }
-        public int CantDivisaAplicaBonoPorPagoDivisa { get; set; }
+        public decimal CantDivisaAplicaBonoPorPagoDivisa { get; set; }
         public decimal MontoBonoPorPagoDivisa { get; set; }
         public decimal MontoBonoEnDivisaPorPagoDivisa { get; set; }
         public decimal MontoPorVueltoEnEfectivo { get; set; }
@@ -33,7 +33,7 @@ namespace PosOnLine.Src.Pago.Procesar
         public void Limpiar() 
         {
             PorctBonoPorPagoDivisa = 0m;
-            CantDivisaAplicaBonoPorPagoDivisa = 0;
+            CantDivisaAplicaBonoPorPagoDivisa = 0m;
             MontoBonoPorPagoDivisa = 0m;
             MontoBonoEnDivisaPorPagoDivisa = 0m;
             MontoPorVueltoEnEfectivo = 0m;
